requirements.txt
Flask==3.0.3
werkzeug==3.0.4
langgraph==1.0.2
langgraph-checkpoint==2.1.0
langgraph-prebuilt==1.0.2
langchain-core==1.2.6  # Latest as of early 2026; satisfies >=1.0.0 for prebuilt
langchain==1.0.7       # Upgrade to 1.x series for compatibility
langchain-openai==1.0.3
langchain-community==0.4.1
langchain-text-splitters==1.0.0
yfinance==0.2.40
pandas==2.2.3
openpyxl==3.1.5
numpy-financial==1.0.0
duckduckgo-search==7.1.1
gunicorn==23.0.0
python-dotenv==1.0.1
pydantic==2.8.2